Featureimage/cmake migration

+++ src/fty_common_db_asset_update.cc
@@ -26,107 +26,85 @@
 @end
 */
 
-#include <tntdb/row.h>
-#include <tntdb/result.h>
-#include <tntdb/error.h>
-
+#include "fty_common_db.h"
 #include <fty_common.h>
-
-#include "fty_common_db_classes.h"
+#include <tntdb/error.h>
+#include <tntdb/result.h>
+#include <tntdb/row.h>
 
 #define ERRCODE_ABNORMAL 1
 
 namespace DBAssetsUpdate {
 
-int
-update_asset_element (tntdb::Connection &conn,
-                      uint32_t element_id,
-                      const char *element_name,
-                      uint32_t parent_id,
-                      const char *status,
-                      uint16_t priority,
-                      const char *asset_tag,
-                      int32_t &affected_rows)
+int update_asset_element(tntdb::Connection& conn, uint32_t element_id, const char* element_name, uint32_t parent_id,
+    const char* status, uint16_t priority, const char* asset_tag, int32_t& affected_rows)
 {
     LOG_START;
-    log_debug ("  element_id = %" PRIi32, element_id);
-    log_debug ("  element_name = '%s'", element_name);
-    log_debug ("  parent_id = %" PRIu32, parent_id);
-    log_debug ("  status = '%s'", status);
-    log_debug ("  priority = %" PRIu16, priority);
-    log_debug ("  asset_tag = '%s'", asset_tag);
+    log_debug("  element_id = %" PRIi32, element_id);
+    log_debug("  element_name = '%s'", element_name);
+    log_debug("  parent_id = %" PRIu32, parent_id);
+    log_debug("  status = '%s'", status);
+    log_debug("  priority = %" PRIu16, priority);
+    log_debug("  asset_tag = '%s'", asset_tag);
 
     // if parent id == 0 ->  it means that there is no parent and value
     // should be updated to NULL
-    try{
+    try {
         tntdb::Statement st = conn.prepareCached(
             " UPDATE"
             "   t_bios_asset_element"
             " SET"
-//            "   name = :name,"
+            //            "   name = :name,"
             "   asset_tag = :asset_tag,"
             "   id_parent = :id_parent,"
             "   status = :status,"
             "   priority = :priority"
-            " WHERE id_asset_element = :id"
-        );
+            " WHERE id_asset_element = :id");
 
-        st = st.set("id", element_id).
-//                           set("name", element_name).
-                           set("status", status).
-                           set("priority", priority).
-                           set("asset_tag", asset_tag);
+        st = st.set("id", element_id)
+                 .
+             //                           set("name", element_name).
+             set("status", status)
+                 .set("priority", priority)
+                 .set("asset_tag", asset_tag);
 
-        if ( parent_id != 0 )
-        {
-            affected_rows = st.set("id_parent", parent_id).
-                               execute();
-        }
-        else
-        {
-            affected_rows = st.setNull("id_parent").
-                               execute();
+        if (parent_id != 0) {
+            affected_rows = int(st.set("id_parent", parent_id).execute());
+        } else {
+            affected_rows = int(st.setNull("id_parent").execute());
         }
         log_debug("[t_asset_element]: updated %" PRIu32 " rows", affected_rows);
         LOG_END;
         // if we are here and affected rows = 0 -> nothing was updated because
         // it was the same
         return 0;
-    }
-    catch (const std::exception &e) {
+    } catch (const std::exception& e) {
         LOG_END_ABNORMAL(e);
         return ERRCODE_ABNORMAL;
     }
 }
 
-int
-update_asset_status_by_name (const char *element_name,
-                            const char *status)
+int update_asset_status_by_name(const char* element_name, const char* status)
 {
     LOG_START;
 
-    if (!streq (status, "active") && !streq (status, "nonactive"))
-    {
-        log_error ("Invalid value of status %s", status);
+    if (!streq(status, "active") && !streq(status, "nonactive")) {
+        log_error("Invalid value of status %s", status);
         return -1;
     }
 
     tntdb::Connection conn = tntdb::connectCached(DBConn::url);
-    tntdb::Statement st = conn.prepareCached(
+    tntdb::Statement  st   = conn.prepareCached(
         " UPDATE"
         "   t_bios_asset_element"
         " SET"
         "   status = :status"
-        " WHERE name = :name"
-    );
+        " WHERE name = :name");
 
-    int32_t affected_rows = st.set("name", element_name).
-                               set("status", status).
-                               execute();
+    int32_t affected_rows = int32_t(st.set("name", element_name).set("status", status).execute());
 
-    if (affected_rows > 1)
-    {
-        log_error ("Name %s should be unique", element_name);
+    if (affected_rows > 1) {
+        log_error("Name %s should be unique", element_name);
         return -1;
     }
 
@@ -135,4 +113,4 @@ update_asset_status_by_name (const char *element_name,
     return 0;
 }
 
-} // namespace end
+} // namespace DBAssetsUpdate
